Move VR controller, fix flag name
