+++ controllers/vulnerabilityreport/vulnerabilityreport_metrics.go
@@ -1,4 +1,4 @@
-package controllers
+package vulnerabilityreport
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
